Matrix.cpp
Nanu went crazy yesterday and created an arbitrary matrix consisting of 0, 1 or 2. There was no rule observed in forming this matrix. But then he made up some rules himself.If there are adjacent 1's, they are said to be connected.Similarly, if there are adjacent 2's, they are said to be connected. Here adjacent means all the surrounding positions along with the diagonal positions.
Now given a matrix of 0, 1 or 2, do your computation according to the following rules: 
1.there is a path from any position in top row to any position in bottom row consisting only of 1, then print 1.
2.If there is a path from any position in first column to any position in last column consisting only of 2, then print 2.
3.If both Rule 1 & Rule 2 are true, print AMBIGUOUSAMBIGUOUS.
4.If none of Rule 1, Rule 2 or Rule 3 satisfies, print 0.